pyproject.toml
[tool.poetry]
name = "flake8_custom_import_rules"
version = "1.1.5"
description = "A Flake8 plugin that enforces custom import rules, allowing users to define and maintain clean and consistent import organization across their Python projects."
readme = "README.rst"
authors = ["Rodrigo Gonzalez <flake8.custom.import.rules@gmail.com>"]
repository = "https://github.com/RodrigoGonzalez/flake8-custom-import-rules"
homepage = "https://github.com/RodrigoGonzalez/flake8-custom-import-rules"
keywords = [
    "code_quality",
    "code_linting",
    "import_rules",
    "flake8",
    "quality_assurance",
]
classifiers = [
    "Framework :: Flake8",
    "Environment :: Console",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]

license = "MIT"

packages = [
    {include = "flake8_custom_import_rules", from = "src"}
]
include = ["README.rst", "LICENSE", "CHANGELOG.md"]

[tool.poetry.scripts]
import-rules = "flake8_custom_import_rules.__main__:main"

[tool.poetry.urls]
"GitHub" = "https://github.com/RodrigoGonzalez/flake8-custom-import-rules"
"Documentation" = "https://RodrigoGonzalez.github.io/flake8-custom-import-rules/"
"Change Log" = "https://github.com/RodrigoGonzalez/flake8-custom-import-rules/blob/main/CHANGELOG.md"

# from https://python-poetry.org/docs/pyproject/#plugins
[tool.poetry.plugins."flake8.extension"]
CIR = "flake8_custom_import_rules.flake8_plugin:Plugin"

[tool.poetry.dependencies]
python = ">=3.9,<=3.12"
flake8 = ">=6.1,<8.0"
importlib-metadata = { version = "*", python = "<3.8" }
pylama = "^8.4.1"
stdlib-list = { version = ">=0.10.0", python = "<3.10" }
pycodestyle = "^2.11.1"
attrs = "^23.2.0"
typer = "^0.9.0"
pymdown-extensions = "^10.7"


[tool.poetry.group.dev.dependencies]
astpretty = "^3.0.0"
ipython = "*"
commitizen = "*"
pre-commit = "*"
isort = "*"
black = "*"
autoflake = "*"
mypy = "*"
lxml = "*"
cmake = "*"
tryceratops = "*"
tomark = "*"


[tool.poetry.group.test.dependencies]
modulegraph = "^0.19.6"
pytest = ">=7.4.4,<9.0.0"
pytest-cov = "^4.1.0"
toml = "^0.10.2"
tox = "^4.12.1"
tox-gh-actions = "^3.2.0"
coverage = { version = "^7.4.1",  extras = ["toml"] }
covdefaults = "^2.3.0"
pytest-mock = "^3.12.0"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.6"
mkdocstrings = {extras = ["optimize", "python"], version = ">=0.22,<0.25"}



[tool.commitizen]
name = "cz_conventional_commits"
version = "1.1.5"
tag_format = "v$version"
version_type = "semver"
update_changelog_on_bump = true
version_files = [
    "pyproject.toml:version",
    "src/flake8_custom_import_rules/__init__.py:__version__",
]
changelog_file = "CHANGELOG.md"

[tool.isort]
profile = "black"
skip_glob = ["docs", "example_repos"]
line_length = 100
multi_line_output = 4
force_single_line = true
src_paths = ["src", "tests", "scripts"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_first_party = ["flake8_custom_import_rules", "my_base_module"]
known_third_party = ["flake8"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
    /(
        \.git
      | \.mypy_cache
      | \.mypy_html_report
      | \.venv
      | _build
      | build
      | dist
      | docs
      | example_repos
    )/
'''

[tool.mypy]
# output
show_error_codes = true
show_column_numbers = true
show_error_context = true
show_traceback = true
# 3rd party import
ignore_missing_imports = true
# dynamic typing
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = false
disallow_subclassing_any = false
# platform
python_version = "3.10"
# untyped
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
# None and Optional
no_implicit_optional = true
# strictness
explicit_package_bases = true
# Warnings
warn_return_any = false
warn_unreachable = true
# Misc
pretty = true
exclude = [
    "example_repos",
    "tests",
]

[tool.coverage.run]
plugins = ["covdefaults"]
omit = ["flake8_custom_import_rules/example_repos/*"]

[tool.tryceratops]
exclude = ["flake8_custom_import_rules/example_repos/", "scripts/*"]
experimental = false
check_pickable = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
# from https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
#requires = ["setuptools", "setuptools-scm"]
#build-backend = "setuptools.build_meta"
